Bug fix for Eudic not coming to the front (#1195)
* Bug fix

Activate Eudic and bring it to the front .

* Bug fix

Activate Eudic and bring it to the front.

* Revert "Bug fix"

This reverts commit 95d0e30c13f3b4be8c3f4ef3341eb41a8370a12f.

* Bug fix

Activate Eudic and bring it to the front.

* Bug fix: Activate Eudic and bring it to the front.

When Eudic runs in the mini bar (with the main window closed), or when minimize the main window into the Dock, the extension doesn't work. In details, PopClip can't activate Eudic and make it to the frontmost.